+++ controllers/index.js
@@ -1,11 +1,13 @@
 const auth = require("./auth");
 const drinks = require("./drinks");
+const mainPopularDrinks = require("./mainPopularDrinks");
 const filters = require("./filters");
 const users = require("./users");
 
 module.exports = {
   auth,
   drinks,
+  mainPopularDrinks,
   filters,
   users,
 };
